pkg/subsystem: move the exiting code to /legacy
It will help gradually switch to the new code, once it's fully ready.
