+++ pkg/subsystem/legacy/maintainers.go
@@ -1,7 +1,7 @@
 // Copyright 2022 syzkaller project authors. All rights reserved.
 // Use of this source code is governed by Apache 2 LICENSE that can be found in the LICENSE file.
 
-package subsystem
+package legacy
 
 func LinuxGetMaintainers(subsystemName string) []string {
 	return linuxSubsystems[subsystemName].cc
